requirements.txt
asgiref==3.5.2
dj-database-url==1.2.0
Django==4.1.4
django-jazzmin==2.6.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
discord_webhook
django-recaptcha